countries v0.1.0 Countries

Provides methods to search countries and access its information.

Functions

country(country_name)

Returns country struct for the given country name.

Examples

iex> Countries.country("india")
%Countries.Country{
  alpha_2: "IN",
  alpha_3: "IND",
  country_code: "356",
  iso_3166_2: "ISO 3166-2:IN",
  name: "India",
  region: "Asia",
  region_code: "142",
  sub_region: "Southern Asia",
  sub_region_code: "034"
}

country_code(country_name)

Returns country_code for the given country name.

Examples

iex> Countries.country_code("india")
"356"

frequency(word)

Returns frequency of word in countries dataset.

Examples

iex> Countries.frequency("Asia")
102

list_countries(region)

Returns list of country names in a given region.

Examples

iex> Countries.list_countries("Oceania")
["American Samoa", "Australia", "Cook Islands", "Fiji", "French Polynesia",
 "Guam", "Kiribati", "Marshall Islands", "Micronesia (Federated States of)",
 "Nauru", "New Caledonia", "New Zealand", "Niue", "Norfolk Island",
 "Northern Mariana Islands", "Palau", "Papua New Guinea", "Pitcairn", "Samoa",
 "Solomon Islands", "Tokelau", "Tonga", "Tuvalu", "Vanuatu",
 "Wallis and Futuna"]